requirements.txt
Django==1.7.1
pinax-theme-bootstrap==5.6.0
biblion==2.2.3pytz==2014.9
django-reversion==1.8.5
